Move geometry coercion to IndexSystem. Update tests.
