+++ src/main/scala/com/databricks/labs/mosaic/core/index/IndexSystem.scala
@@ -2,7 +2,8 @@ package com.databricks.labs.mosaic.core.index
 
 import com.databricks.labs.mosaic.core.geometry.MosaicGeometry
 import com.databricks.labs.mosaic.core.geometry.api.GeometryAPI
-import com.databricks.labs.mosaic.core.types.model.MosaicChip
+import com.databricks.labs.mosaic.core.types.model.{GeometryTypeEnum, MosaicChip}
+import com.databricks.labs.mosaic.core.types.model.GeometryTypeEnum._
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.UTF8String
 
@@ -22,15 +23,16 @@ abstract class IndexSystem(var cellIdType: DataType) extends Serializable {
 
     def getResolutionStr(resolution: Int): String
 
-    def formatCellId(cellId: Any, dt: DataType): Any = (dt, cellId) match {
-        case (LongType, _: Long)           => cellId
-        case (LongType, cid: String)       => parse(cid)
-        case (LongType, cid: UTF8String)   => parse(cid.toString)
-        case (StringType, cid: Long)       => format(cid)
-        case (StringType, cid: UTF8String) => cid.toString
-        case (StringType, _: String)       => cellId
-        case _                             => throw new Error("Cell ID data type not supported.")
-    }
+    def formatCellId(cellId: Any, dt: DataType): Any =
+        (dt, cellId) match {
+            case (LongType, _: Long)           => cellId
+            case (LongType, cid: String)       => parse(cid)
+            case (LongType, cid: UTF8String)   => parse(cid.toString)
+            case (StringType, cid: Long)       => format(cid)
+            case (StringType, cid: UTF8String) => cid.toString
+            case (StringType, _: String)       => cellId
+            case _                             => throw new Error("Cell ID data type not supported.")
+        }
 
     def formatCellId(cellId: Any): Any = formatCellId(cellId, getCellIdDataType)
 
@@ -158,9 +160,10 @@ abstract class IndexSystem(var cellIdType: DataType) extends Serializable {
         val intersections = for (index <- borderIndices) yield {
             val indexGeom = indexToGeometry(index, geometryAPI)
             val intersect = geometry.intersection(indexGeom)
-            val isCore = intersect.equals(indexGeom)
+            val coerced = coerceChipGeometry(intersect)
+            val isCore = coerced.equals(indexGeom)
 
-            val chipGeom = if (!isCore || keepCoreGeom) intersect else null
+            val chipGeom = if (!isCore || keepCoreGeom) coerced else null
 
             MosaicChip(isCore = isCore, Left(index), chipGeom)
         }
@@ -218,5 +221,29 @@ abstract class IndexSystem(var cellIdType: DataType) extends Serializable {
       *   Index ID in this index system.
       */
     def pointToIndex(lon: Double, lat: Double, resolution: Int): Long
+
+    def coerceChipGeometry(geom: MosaicGeometry): MosaicGeometry = {
+        val geomType = GeometryTypeEnum.fromString(geom.getGeometryType)
+        if (geomType == GEOMETRYCOLLECTION) {
+            val geometries = geom.flatten
+            val coerced = coerceChipGeometry(geometries)
+            coerced.reduce(_ union _)
+        } else {
+            geom
+        }
+    }
+
+    def coerceChipGeometry(geometries: Seq[MosaicGeometry]): Seq[MosaicGeometry] = {
+        val types = geometries.map(_.getGeometryType).map(GeometryTypeEnum.fromString)
+        if (types.contains(MULTIPOLYGON) || types.contains(POLYGON)) {
+            geometries.filter(g => Seq(POLYGON, MULTIPOLYGON).contains(GeometryTypeEnum.fromString(g.getGeometryType)))
+        } else if (types.contains(MULTILINESTRING) || types.contains(LINESTRING)) {
+            geometries.filter(g => Seq(MULTILINESTRING, LINESTRING).contains(GeometryTypeEnum.fromString(g.getGeometryType)))
+        } else if (types.contains(MULTIPOINT) || types.contains(POINT)) {
+            geometries.filter(g => Seq(MULTIPOINT, POINT).contains(GeometryTypeEnum.fromString(g.getGeometryType)))
+        } else {
+            Nil
+        }
+    }
 
 }
